New theme: Nobel Prize Winners
December 10 is the day Nobel Prizes are awarded, so it may be
fitting to release this theme on December 11.

Since (AFAIK) you can't use a RemoteList for a theme that has
categories, nobel_prize.pm is generated by a script. And since
MultiList::init doesn't take arguments, the generated module
uses some hackery to populate the categories from the POD.